utility shared function file
